# Rewrite boostable bosses parser